Finish update platform build tags
