--- receiver/dcgmreceiver/factory_windows.go
@@ -1,36 +1,0 @@
-// Copyright 2023 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build windows
-// +build windows
-
-package dcgmreceiver
-
-import (
-	"context"
-	"errors"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/receiver"
-)
-
-func createMetricsReceiver(
-	_ context.Context,
-	params receiver.CreateSettings,
-	rConf component.Config,
-	consumer consumer.Metrics,
-) (receiver.Metrics, error) {
-	return nil, errors.New("DCGM receiver is only supported on Linux")
-}
