+++ receiver/nvmlreceiver/scraper.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !windows
-// +build !windows
+//go:build linux
+// +build linux
 
 package nvmlreceiver
